+++ packages/frontend-api/src/Model/Mutation/Customer/DeliveryAddress/Exception/DeliveryAddressNotFoundUserError.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace App\FrontendApi\Mutation\Customer\Exception;
+namespace Shopsys\FrontendApiBundle\Model\Mutation\Customer\DeliveryAddress\Exception;
 
 use Shopsys\FrontendApiBundle\Model\Error\EntityNotFoundUserError;
 use Shopsys\FrontendApiBundle\Model\Error\UserErrorWithCodeInterface;
 
 class DeliveryAddressNotFoundUserError extends EntityNotFoundUserError implements UserErrorWithCodeInterface
 {
-    private const CODE = 'delivery-address-not-found';
+    protected const CODE = 'delivery-address-not-found';
 
     /**
      * {@inheritdoc}
      */
     public function getUserErrorCode(): string
     {
-        return self::CODE;
+        return static::CODE;
     }
 }
