+++ project-base/app/src/Model/Customer/DeliveryAddressDataFactory.php
@@ -10,6 +10,7 @@ use Shopsys\FrameworkBundle\Model\Customer\DeliveryAddressDataFactory as BaseDel
  * @method \App\Model\Customer\DeliveryAddressData create()
  * @method \App\Model\Customer\DeliveryAddressData createFromDeliveryAddress(\App\Model\Customer\DeliveryAddress $deliveryAddress)
  * @method fillFromDeliveryAddress(\App\Model\Customer\DeliveryAddressData $deliveryAddressData, \App\Model\Customer\DeliveryAddress $deliveryAddress)
+ * @method \App\Model\Customer\DeliveryAddressData createFromDeliveryInputArgumentAndCustomer(\Overblog\GraphQLBundle\Definition\Argument $argument, \Shopsys\FrameworkBundle\Model\Customer\Customer $customer)
  */
 class DeliveryAddressDataFactory extends BaseDeliveryAddressDataFactory
 {
